Move mat_double from powell_lsq.c to utils.c
