+++ src/utils.h
@@ -42,6 +42,8 @@ void power_m(int count, double* result, const double* base,
 
 char* fgets_potfit(char* buffer, int len, FILE* f);
 
+double** mat_double(int, int);
+
 static inline int min(int a, int b) { return a < b ? a : b; }
 static inline int max(int a, int b) { return a > b ? a : b; }
